+++ src/instances/garbage/instance.ts
@@ -1,7 +1,7 @@
 import type { WorkbenchInstanceRegistration } from '../instance-types'
 import { garbageInstance } from './config'
 
-export const garbageWorkbenchInstance = {
+export const workbenchInstanceRegistration = {
   instance: garbageInstance,
   default: true,
 } satisfies WorkbenchInstanceRegistration