+++ src/instances/greathouse/instance.ts
@@ -2,7 +2,7 @@ import type { WorkbenchInstanceRegistration } from '../instance-types'
 import { greathouseInstance } from './config'
 import { greathousePilotSnapshot } from './pilot'
 
-export const greathouseWorkbenchInstance = {
+export const workbenchInstanceRegistration = {
   instance: greathouseInstance,
   staticSnapshot: greathousePilotSnapshot,
 } satisfies WorkbenchInstanceRegistration